Updated PHP and Form data.
This will give us a form to submit information to.  I wanted to create the form data using my own code, but this turned out to be a problem because I would also need to setup a mail server in my raspbery pi as well. Instead I opted to use a Google form instead.